requirements.txt
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
toml==0.10.2
